Don't change patch state after running job
A runner client queries patchwork site to check what jobs need to be
done. The query was done using:
"/api/events/?category=patch-completed"

Before this commit, after the job was finished the state was changed so
that the query will not return the same patch. However this approach
will not work when there may be multiple runners. Instead we shouldn't
change the category but filter by date (i.e. number of new patches in
the last 2 hours). Now we also need to keep track locally what patches
we have processed so that we don't run the same job twice (use a
separate table for this).

Signed-off-by: Andriy Gelman <andriy.gelman@gmail.com>