Fetch patches over a variable time interval
When querying for new patches a time interval must be specified (i.e.
fetch ready patches received over the last 2 hours). Currently this
fetch interval is fixed. However when processing a large patch set the
run time for CI can be significantly longer. Thus new patches may be
missed in the next query. To fix this issue make the time interval
a variable and set it by measuring the execution time of the CI job.

Signed-off-by: Andriy Gelman <andriy.gelman@gmail.com>